@@ -10,17 +10,17 @@
         }
     ],
     "require": {
-        "php": ">=7.4",
+        "php": "^8.0 || ^8.1 || ^8.2",
         "ext-json": "*",
         "cebe/php-openapi": "^1.7",
-        "fakerphp/faker": "^1.19",
+        "fakerphp/faker": "^1.20",
         "league/openapi-psr7-validator": "^0.18",
-        "thecodingmachine/safe": "^1.3"
+        "thecodingmachine/safe": "^2.4"
     },
     "require-dev": {
-        "doctrine/coding-standard": "^9",
+        "doctrine/coding-standard": "^10",
         "ergebnis/composer-normalize": "^2.27",
-        "infection/infection": "^0.25",
+        "infection/infection": "^0.26",
         "phpstan/extension-installer": "^1.1",
         "phpstan/phpstan": "^1.7",
         "phpstan/phpstan-phpunit": "^1",
@@ -45,7 +45,8 @@
             "dealerdirect/phpcodesniffer-composer-installer": true,
             "infection/extension-installer": true,
             "ergebnis/composer-normalize": true,
-            "phpstan/extension-installer": true
+            "phpstan/extension-installer": true,
+            "ocramius/package-versions": true
         },
         "sort-packages": true
     },
