@@ -28,7 +28,7 @@ class BooleanFakerTest extends UnitTestCase
     /** @test */
     function it_can_generate_boolean_value()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: boolean
 YAML;
 
@@ -40,7 +40,7 @@ YAML;
     /** @test */
     function it_can_generate_boolean_value_from_enum()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: boolean
 enum:
   - true
