@@ -28,7 +28,7 @@ class NumberFakerTest extends UnitTestCase
     /** @test */
     function it_can_generate_a_number()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 YAML;
 
@@ -41,7 +41,7 @@ YAML;
     /** @test */
     function it_can_handle_number_float_format()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 format: float
 YAML;
@@ -55,7 +55,7 @@ YAML;
     /** @test */
     function it_can_handle_number_double_format()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 format: double
 YAML;
@@ -69,7 +69,7 @@ YAML;
     /** @test */
     function it_can_generate_a_integer()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 YAML;
 
@@ -82,7 +82,7 @@ YAML;
     /** @test */
     function it_can_handle_integer_int32_format()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 format: int32
 YAML;
@@ -96,7 +96,7 @@ YAML;
     /** @test */
     function it_can_handle_integer_int64_format()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 format: int64
 YAML;
@@ -110,7 +110,7 @@ YAML;
     /** @test */
     function it_can_handle_minimum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 minimum: 100
 YAML;
@@ -125,7 +125,7 @@ YAML;
     /** @test */
     function it_can_handle_maximum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 maximum: 100
 YAML;
@@ -140,7 +140,7 @@ YAML;
     /** @test */
     function it_can_handle_both_minimum_and_maximum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 minimum: 100
 maximum: 200
@@ -155,7 +155,7 @@ YAML;
     /** @test */
     function it_can_handle_exclusive_minimum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 minimum: 100
 maximum: 102
@@ -172,7 +172,7 @@ YAML;
     /** @test */
     function it_can_handle_exclusive_maximum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 minimum: 100
 maximum: 102
@@ -189,7 +189,7 @@ YAML;
     /** @test */
     function it_can_handle_both_exclusive_minimum_and_exclusive_maximum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 minimum: 100
 maximum: 102
@@ -207,7 +207,7 @@ YAML;
     /** @test */
     function it_can_handle_multiple_of_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 multipleOf: 10
 YAML;
@@ -221,7 +221,7 @@ YAML;
     /** @test */
     function it_can_handle_multiple_of_keyword_with_float_number()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 multipleOf: 8
 YAML;
@@ -235,7 +235,7 @@ YAML;
     /** @test */
     function it_can_generate_elements_from_enum()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 enum:
   - 1010.9865
