@@ -29,7 +29,7 @@ class StaticNumberFakerTest extends UnitTestCase
     /** @test */
     function it_can_generate_a_number()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 YAML;
 
@@ -42,7 +42,7 @@ YAML;
     /** @test */
     function it_can_handle_number_float_format()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 format: float
 YAML;
@@ -56,7 +56,7 @@ YAML;
     /** @test */
     function it_can_handle_number_double_format()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 format: double
 YAML;
@@ -70,7 +70,7 @@ YAML;
     /** @test */
     function it_can_generate_a_integer()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 YAML;
 
@@ -83,7 +83,7 @@ YAML;
     /** @test */
     function it_can_handle_integer_int32_format()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 format: int32
 YAML;
@@ -97,7 +97,7 @@ YAML;
     /** @test */
     function it_can_handle_integer_int64_format()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 format: int64
 YAML;
@@ -111,7 +111,7 @@ YAML;
     /** @test */
     function it_will_return_a_integer_if_unknown_format_is_given()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 format: unkown
 YAML;
@@ -125,7 +125,7 @@ YAML;
     /** @test */
     function it_can_handle_minimum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 minimum: 100
 YAML;
@@ -140,7 +140,7 @@ YAML;
     /** @test */
     function it_can_handle_maximum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 maximum: 100
 YAML;
@@ -155,7 +155,7 @@ YAML;
     /** @test */
     function it_can_handle_both_minimum_and_maximum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 minimum: 100
 maximum: 200
@@ -171,7 +171,7 @@ YAML;
     /** @test */
     function it_can_handle_exclusive_minimum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 minimum: 100
 maximum: 102
@@ -188,7 +188,7 @@ YAML;
     /** @test */
     function it_can_handle_exclusive_maximum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 minimum: 100
 maximum: 102
@@ -206,7 +206,7 @@ YAML;
     /** @test */
     function it_can_handle_both_exclusive_minimum_and_exclusive_maximum_keyword()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 minimum: 100
 maximum: 102
@@ -224,7 +224,7 @@ YAML;
     /** @test */
     function it_can_handle_multiple_of_keyword_with_integer_number()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: integer
 multipleOf: 10
 YAML;
@@ -238,7 +238,7 @@ YAML;
     /** @test */
     function it_can_handle_multiple_of_keyword_with_float_number()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 multipleOf: 8
 YAML;
@@ -252,7 +252,7 @@ YAML;
     /** @test */
     function it_can_generate_elements_from_enum()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 enum:
   - 1010.9865
@@ -269,7 +269,7 @@ YAML;
     /** @test */
     function it_can_generate_default_value_from_enum()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 type: number
 enum:
     - 1010.9865
@@ -286,7 +286,7 @@ YAML;
     /** @test */
     function it_can_generate_nullable_value()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
  type: number
  nullable: true
  YAML;
@@ -299,7 +299,7 @@ YAML;
     /** @test */
     function it_can_generate_example_value()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
  type: number
  example: 10.33
  YAML;
