@@ -36,7 +36,7 @@ class StaticRequestFakerTest extends UnitTestCase
     /** @test */
     function it_will_mock_the_request()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 schema:
   type: object
   required:
@@ -60,7 +60,7 @@ YAML;
     /** @test */
     function it_will_mock_the_first_example()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 schema:
   type: object
   required:
@@ -97,7 +97,7 @@ YAML;
     /** @test */
     function it_will_mock_the_given_example()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 schema:
   type: object
   required:
@@ -134,7 +134,7 @@ YAML;
     /** @test */
     function it_throws_exception_if_example_cannot_be_found()
     {
-        $yaml = <<<YAML
+        $yaml = <<<'YAML'
 schema:
   type: object
   required:
