@@ -7,9 +7,7 @@ namespace Vural\OpenAPIFaker\Tests\Unit\Utils;
 use Vural\OpenAPIFaker\Tests\Unit\UnitTestCase;
 use Vural\OpenAPIFaker\Utils\RegexUtils;
 
-/**
- * @covers \Vural\OpenAPIFaker\Utils\RegexUtils
- */
+/** @covers \Vural\OpenAPIFaker\Utils\RegexUtils */
 class RegexUtilsTest extends UnitTestCase
 {
     private const SAMPLE_REGEX = '^[a-z]-[0-9]-A{2}-[ABC]{1,2}-(foo|bar){1,2}-(A)+-(\d{3}-\d{2}-\d{4})*.*$';
